Move WineryConnector to control modul
