+++ org.opentosca.container.api/pom.xml
@@ -25,11 +25,6 @@
 		</dependency>
 		<dependency>
 			<groupId>org.opentosca</groupId>
-			<artifactId>org.opentosca.container.connector.winery</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.opentosca</groupId>
 			<artifactId>org.opentosca.planbuilder.core</artifactId>
 			<version>${project.version}</version>
 		</dependency>
