+++ org.opentosca.container.control/pom.xml
@@ -1,38 +1,53 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
          xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
+    <modelVersion>4.0.0</modelVersion>
 
-  <parent>
-    <groupId>org.opentosca</groupId>
-    <artifactId>container</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-  </parent>
+    <parent>
+        <groupId>org.opentosca</groupId>
+        <artifactId>container</artifactId>
+        <version>3.0.0-SNAPSHOT</version>
+    </parent>
 
-  <artifactId>org.opentosca.container.control</artifactId>
+    <artifactId>org.opentosca.container.control</artifactId>
 
-  <dependencies>
-    <dependency>
-      <groupId>org.opentosca</groupId>
-      <artifactId>org.opentosca.container.core</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.opentosca</groupId>
-      <artifactId>org.opentosca.container.plan.deployment.core</artifactId>
-      <version>${project.version}</version>
-    </dependency>
+    <dependencies>
+        <dependency>
+            <groupId>org.opentosca</groupId>
+            <artifactId>org.opentosca.container.core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.opentosca</groupId>
+            <artifactId>org.opentosca.container.plan.deployment.core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
+            <version>4.5.13</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpmime</artifactId>
+            <version>4.5.1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+            <version>2.9.10.8</version>
+        </dependency>
 
-    <dependency>
-      <groupId>com.github.opentosca.winery</groupId>
-      <artifactId>org.eclipse.winery.model.tosca.canonical</artifactId>
-      <version>ustutt-SNAPSHOT</version>
-    </dependency>
-      <dependency>
-          <groupId>org.opentosca</groupId>
-          <artifactId>org.opentosca.planbuilder.integration</artifactId>
-          <version>3.0.0-SNAPSHOT</version>
-          <scope>compile</scope>
-      </dependency>
-  </dependencies>
+        <dependency>
+            <groupId>com.github.opentosca.winery</groupId>
+            <artifactId>org.eclipse.winery.model.tosca.canonical</artifactId>
+            <version>ustutt-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.opentosca</groupId>
+            <artifactId>org.opentosca.planbuilder.integration</artifactId>
+            <version>3.0.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+    </dependencies>
 </project>
